+++ src/Yoti.Auth/DigitalIdentity/ShareSessionResult.cs
@@ -5,10 +5,9 @@ using Yoti.Auth.DataObjects;
 
 namespace Yoti.Auth.DigitalIdentity
 {
-    public class ShareSessionResult
+    public class GetSessionResult
     {
 #pragma warning disable 0649
-
         // These fields are assigned to by JSON deserialization
         [JsonProperty("id")]
         public string Id { get; set; }
@@ -19,31 +18,9 @@ namespace Yoti.Auth.DigitalIdentity
         [JsonProperty("expiry")]
         public string Expiry { get; set; }
 
-        [JsonProperty("created")]
-        public string Created { get; set; }
-
-        [JsonProperty("updated")]
-        public string Updated { get; set; }
-
-        [JsonProperty("qrCode")]
-        public qrCode QrCode { get; set; }
-
-        [JsonProperty("receipt")]
-        public receipt Receipt { get; set; }
-
 #pragma warning restore 0649
 
     }
 
-    public class qrCode
-    {
-        [JsonProperty("id")]
-        public string Id { get; set; }
-    }
-
-    public class receipt
-    {
-        [JsonProperty("id")]
-        public string Id { get; set; }
-    }
+    
 }
